requirements.txt
# OBIE - OSINT Buyer Intent Engine
# Core Dependencies

# Browser automation
playwright>=1.40.0
playwright-stealth>=1.0.6

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# DNS and email validation
dnspython>=2.4.0

# LLM integration (Groq for fast, free-tier inference)
groq>=0.4.0

# Retry logic
tenacity>=8.2.0

# Data processing
python-dateutil>=2.8.0
